Rename libs/cvr-fixture-generator to libs/fixture-generators
Make room for adding an election fixture generator alongside the CVR
fixture generator
