+++ libs/fixture-generators/README.md
@@ -1,14 +1,17 @@
-# Cast Vote Record Fixture Generator
+# Fixture Generators
+
+Tools for generating election data fixtures for testing.
+
+## Cast Vote Record Fixture Generator
 
 A command-line tool for generating cast vote record fixtures from ballot
 packages. Generates BMD ballot CVRs for VotingWorks election definitions and
 HMPB ballot CVRs for NH (`gridLayouts`) election definitions
 
-## Usage
+### Usage
 
 ```bash
-# from libs/cvr-fixture-generator
-./bin/generate --electionDefinition ./election.json --outputPath ./standard-cvr-fixture
+./bin/generate-cvrs --electionDefinition ./election.json --outputPath ./standard-cvr-fixture
 ```
 
 Optional flags:
@@ -26,15 +29,15 @@ Optional flags:
   the CDF). Since ballot ids are incrementing numbers starting from 1, they will
   have id collisions in VxAdmin if they are not distinguished by a prefix.
 
-## Saved Fixtures
+### Saved Fixtures
 
 To regenerate the saved fixtures in [libs/fixtures](../libs/fixtures), run:
 
 ```bash
-pnpm generate-fixtures
+pnpm generate-cvr-fixtures
 ```
 
-## Vote Variations
+### Vote Variations
 
 The export may include the following vote variations for candidate contests:
 
@@ -47,7 +50,7 @@ The export may include the following vote variations for candidate contests:
 For ballot measure contests, each of the four possible vote variations are
 included.
 
-## Limitations
+### Limitations
 
 - Multi-sheet ballots are not supported. There is no underlying technical
   limitation, we just do not have the fixtures or the requirement yet to test
