+++ libs/fixture-generators/bin/generate-cvr-fixtures
