+++ libs/fixture-generators/src/cli/generate-cvrs/index.ts
