+++ libs/fixture-generators/src/generate-cvrs/index.ts
@@ -1,1 +1,2 @@
 export * from './generate_cvrs';
+export * from './utils';
