+++ libs/fixture-generators/src/index.ts
@@ -1,0 +1,1 @@
+export * from './generate-cvrs/generate_cvrs';
