+++ libs/fixtures/src/data/electionGridLayoutNewHampshireTestBallot/index.ts
@@ -32,7 +32,7 @@ export * as templateFront from './template-front.jpeg';
 export * as templateBack from './template-back.jpeg';
 export * as templatePdf from './template.pdf';
 
-// Generated by libs/cvr-fixture-generator script: pnpm generate-fixtures
+// Generated by libs/fixture-generators script: pnpm generate-cvr-fixtures
 export const castVoteRecordExport = {
   asDirectoryPath: () =>
     path.join(castVoteRecords.asDirectoryPath(), 'generated'),
