+++ libs/fixtures/src/data/electionTwoPartyPrimary/index.ts
@@ -27,7 +27,7 @@ export const singlePrecinctElectionDefinition = asElectionDefinition(
 
 export * as systemSettings from '../systemSettings.json';
 
-// Generated by libs/cvr-fixture-generator script: pnpm generate-fixtures
+// Generated by libs/fixture-generators script: pnpm generate-cvr-fixtures
 export const castVoteRecordExport = {
   asDirectoryPath: () =>
     path.join(castVoteRecords.asDirectoryPath(), 'generated'),
